chore(project): reorganize folders and decouple logic from UI
               - Rename `entities/Tags` → `entities/tag`, `entities/Category` → `entities/category` (use singular)
               - Move `useCategory` and `useTags` hooks from widgets into `shared/lib/hooks`
               - Restructure `widgets/Post`:
                 - Split into `ui/` (PostCard, PostMeta, PostActions) and `form/` (PostForm, PostFormFields, PostFormOptions)
               - Extract pure Auth form UI into `widgets/AuthForm` and leave container logic in `features/Auth`
               - Remove unnecessary `lib` folder under `widgets/Post/form` and relocate its `types.ts` directly into `form/`
