--- src/shared/config/appwrite/index.ts
@@ -1,8 +1,0 @@
-// import { Account, Client, Databases } from 'appwrite';
-//
-// const client = new Client()
-//   .setEndpoint(import.meta.env.VITE_APP_ENDPOINT || '')
-//   .setProject(import.meta.env.VITE_APP_PROJECT_ID || '');
-
-// export const account = new Account(client);
-// export const db = new Databases(client);
