+++ src/shared/lib/hooks/useCategory.ts
